IcePassive: - moved class to subdir to mirror core pk's filestructure - refactored result caching to use SkateConfig for better readability and maintainability
