+++ src/com/jedk1/jedcore/JCMethods.java
@@ -2,6 +2,7 @@ package com.jedk1.jedcore;
 
 import java.util.*;
 
+import com.jedk1.jedcore.ability.waterbending.passive.IcePassive;
 import com.jedk1.jedcore.util.*;
 import com.projectkorra.projectkorra.ability.ElementalAbility;
 import com.projectkorra.projectkorra.ability.WaterAbility;
@@ -352,5 +353,6 @@ public class JCMethods {
 		RegenTempBlock.revertAll();
 		JedCore.plugin.initializeCollisions();
 		FireTick.loadMethod();
+		IcePassive.loadConfig();
 	}
 }
